# BUGFIX - type mismatch of timestamp (#2)
* Timestamp는 10진수 문자열입니다. 초단위 시간을 표현합니다.
* 하지만 Integer로 저장하고 데이터를 주고받고 있었습니다.

* integer로 저장하던 부분을 모두 String으로 수정한 뒤, 필요에 따라 integer type converting을 합니다.

+++ app/src/main/java/io/tethys/tethyswallet/data/grpc/message/response/MsgUnpacker.kt
@@ -66,7 +66,7 @@ class MsgUnpacker(
     private fun checkSender(): Boolean = body?.let { header.sender.equals(body.merger) } ?: false
 
     private fun checkTime(): Boolean =
-        body?.let { (body.time + MSG_EXP_TIME) > getTimestamp() } ?: false
+        body?.let { (body.time.toInt() + MSG_EXP_TIME) > getTimestamp() } ?: false
 
     fun checkValidity(): Reply.Status {
         if (!checkTime()) return Reply.Status.ECDH_TIMEOUT
